FxS8のDivの実装を修正

+++ src/types.rs
@@ -53,8 +53,8 @@ impl Div for FxS8 {
     fn div(self,other:FxS8) -> FxS8 {
         (((((self.raw as u32) << 3) &
             0b01111111 |
-            ((self.raw as u32) & 0b10000000)) as i32 /
-            (other.raw as i32)) as i8).into()
+            ((self.raw as u32) & 0b10000000)) /
+            (other.raw as u32)) as i8).into()
     }
 }
 impl Neg for FxS8 {
